Add support for ANDROID_TZDATA_ROOT
Soon, code in the Android Core Library / ICU4C library will depend on the
ANDROID_TZDATA_ROOT environment variable being set. On device this points
to the root of the com.android.tzdata module.

Unlike ANDROID_RUNTIME_ROOT, ANDROID_TZDATA_ROOT it doesn't have to point to
anything that exists because the code will fall back to looking in
ANDROID_RUNTIME_ROOT if it can't find the files it is looking for. Most
of the changes in this commit are plumbing to make sure the environment
variable is set to benign defaults or passed through to the runtime.

This commit has been put together by looking for usages of
ANDROID_RUNTIME_ROOT and duplicating the logic.

Bug: 128422035
Test: build only
Change-Id: I2fd73fe16f5022742aaf634be158765cac8759e6